Pull camera functionality of map view into separate class.
