+++ app/src/main/java/org/opus/beacon/MapActivity.java
@@ -10,13 +10,9 @@ import android.location.LocationListener;
 import android.os.AsyncTask;
 import android.os.Handler;
 import android.support.v4.app.ActivityCompat;
-import android.support.v4.app.FragmentActivity;
 import android.os.Bundle;
 
-import com.google.android.gms.maps.CameraUpdate;
-import com.google.android.gms.maps.CameraUpdateFactory;
 import com.google.android.gms.maps.GoogleMap;
-import com.google.android.gms.maps.SupportMapFragment;
 import com.google.android.gms.maps.model.BitmapDescriptorFactory;
 import com.google.android.gms.maps.model.CameraPosition;
 import com.google.android.gms.maps.model.LatLng;
@@ -32,29 +28,19 @@ import android.widget.Toast;
 import java.util.ArrayList;
 import java.util.HashMap;
 
-public class MapActivity extends FragmentActivity
+public class MapActivity extends ConstrainedCameraMap
     implements LocationListener,
     ActivityCompat.OnRequestPermissionsResultCallback,
-    GoogleMap.OnMarkerClickListener,
-    GoogleMap.OnCameraChangeListener {
+    GoogleMap.OnMarkerClickListener {
 
-    private GoogleMap mMap;
     private LocationManager mLocationManager;
     private Context context;
 
     private float GPS_ACCURACY = 50.0f; //meters
     private long GPS_WAIT = 4000; //milliseconds
 
-    private float MAX_ZOOM = 18.0f;
-    private float MIN_ZOOM = 3.0f;
-
-    private float MAX_TILT = 60.0f;
-    private float MIN_TILT = 0.0f;
-
     private int ACCESS_FINE_LOCATION_TAG = 125;
     private int BEACON_SUBMISSION = 37;
-
-    private boolean performedInitialZoom = false;
 
     private BeaconRestClient mClient;
     private Auth mAuth;
@@ -66,8 +52,6 @@ public class MapActivity extends FragmentActivity
         super.onCreate(savedInstanceState);
         context = this;
         mMarkerHash = new HashMap<Marker, BeaconThumb>();
-        setContentView(R.layout.activity_map);
-        setUpMapIfNeeded();
 
         try {
             mAuth = new Auth(this);
@@ -78,52 +62,17 @@ public class MapActivity extends FragmentActivity
     }
 
     @Override
-    protected void onResume() {
-        super.onResume();
-        setUpMapIfNeeded();
-    }
-
-
-    @Override
     protected void onActivityResult(int requestCode, int resultCode, Intent data) {
         if (requestCode == BEACON_SUBMISSION && resultCode == RESULT_OK) {
            localize();
         }
     }
 
-    private void setUpMapIfNeeded() {
-        // Do a null check to confirm that we have not already instantiated the map.
-        if (mMap == null) {
-            // Try to obtain the map from the SupportMapFragment.
-            mMap = ((SupportMapFragment) getSupportFragmentManager().findFragmentById(R.id.map))
-                    .getMap();
-            // Check if we were successful in obtaining the map.
-            if (mMap != null) {
-                setUpMap();
-            }
-        }
-    }
-
-    private void setUpMap() {
-        mMap.setOnMarkerClickListener(this);
-        mMap.setOnCameraChangeListener(this);
-        mMap.getUiSettings().setRotateGesturesEnabled(false);
-        mMap.getUiSettings().setTiltGesturesEnabled(false);
-        mMap.getUiSettings().setCompassEnabled(false);
-        mMap.getUiSettings().setMapToolbarEnabled(false);
-        mMap.getUiSettings().setMyLocationButtonEnabled(false);
-        mMap.getUiSettings().setIndoorLevelPickerEnabled(false);
-
-        // start at center of contiguous United States
-        // this location should probably be a project-level property
-        CameraUpdate center = CameraUpdateFactory.newLatLngZoom(new LatLng(39.8282f, -98.5795f), MIN_ZOOM);
-
-        if (!performedInitialZoom) {
-            mMap.moveCamera(center);
-        }
-
+    @Override
+    protected void setUpMap() {
+        super.setUpMap();
         mLocationManager = (LocationManager) getSystemService(Context.LOCATION_SERVICE);
-
+        mMap.setOnMarkerClickListener(this);
         localize();
     }
 
@@ -150,41 +99,6 @@ public class MapActivity extends FragmentActivity
 
     private boolean shouldUseLocation(Location location) {
         return location.getAccuracy() < GPS_ACCURACY;
-    }
-
-    private void zoomToLocation(Location location) {
-        float zoom;
-        if (!performedInitialZoom) {
-            zoom = MAX_ZOOM;
-            performedInitialZoom = true;
-        } else {
-            zoom = mCurrentCamera.zoom;
-        }
-
-        LatLng newLocation = new LatLng(location.getLatitude(), location.getLongitude());
-        if (mCurrentCamera != null) {
-            LatLng lastLocation = mCurrentCamera.target;
-
-            float[] results = new float[1];
-            Location.distanceBetween(newLocation.latitude,
-                    newLocation.longitude,
-                    lastLocation.latitude,
-                    lastLocation.longitude,
-                    results);
-            float dist = results[0];
-
-            if (dist < 30.0f) {
-                newLocation = lastLocation;
-            }
-        }
-
-        CameraPosition movement = new CameraPosition.Builder()
-                .target(newLocation)
-                .zoom(zoom)
-                .tilt(MAX_TILT)
-                .build();
-
-        mMap.animateCamera(CameraUpdateFactory.newCameraPosition(movement), 2000, null);
     }
 
     private void performInitialZoom(Location location) {
@@ -216,19 +130,13 @@ public class MapActivity extends FragmentActivity
     }
 
     @Override
-    public void onStatusChanged(String provider, int status, Bundle extras) {
-
-    }
+    public void onStatusChanged(String provider, int status, Bundle extras) {}
 
     @Override
-    public void onProviderEnabled(String provider) {
-
-    }
+    public void onProviderEnabled(String provider) {}
 
     @Override
-    public void onProviderDisabled(String provider) {
-
-    }
+    public void onProviderDisabled(String provider) {}
 
     @Override
     public void onRequestPermissionsResult(int requestCode, String permissions[], int[] grantResults) {
@@ -249,97 +157,28 @@ public class MapActivity extends FragmentActivity
         return false;
     }
 
-    @Override
-    public void onCameraChange(CameraPosition position) {
-        mCurrentCamera = position;
-        loadNewBeacons(position);
-        mCameraHandler.removeCallbacks(mCameraCheckRunnable);
-        checkCameraAnimating();
-    }
-
     private Location mLastLoad = null;
     private void loadNewBeacons(CameraPosition position) {
         if (mLastLoad == null)
             return;
 
-        float[] results = new float[1];
-        Location.distanceBetween(mLastLoad.getLatitude(),
-                mLastLoad.getLongitude(),
-                position.target.latitude,
-                position.target.longitude,
-                results);
-        float dist = results[0];
+        float dist = distanceBetween(mLastLoad, position.target);
 
         // distance greater than a mile
         Location newLocation = new Location("");
         newLocation.setLatitude(position.target.latitude);
         newLocation.setLongitude(position.target.longitude);
 
-        double zoomThreshold = (MAX_ZOOM + MIN_ZOOM) * 0.5f;
-        if (dist > 1609.0f && position.zoom > zoomThreshold) {
+        float zoom = getZoomRatio(position);
+        if (dist > 1609.0f && zoom > 0.5f) {
             new GetLocalBeacons().execute(newLocation);
         }
     }
 
-    public void onCameraAnimationEnd() {
-        float curZoom = mCurrentCamera.zoom;
-        LatLng location = mCurrentCamera.target;
-        float curTilt;
-
-        if (curZoom > MAX_ZOOM) {
-            curTilt = MAX_TILT;
-        } else if (curZoom < MIN_ZOOM) {
-            curTilt = MIN_TILT;
-        } else {
-            float m = (MAX_TILT - MIN_TILT) / (MAX_ZOOM - MIN_ZOOM);
-            curTilt = m * (curZoom - MIN_ZOOM) + MIN_TILT;
-        }
-
-        final CameraPosition movement = new CameraPosition.Builder()
-                .target(location)
-                .zoom(curZoom)
-                .tilt(curTilt)
-                .build();
-
-        runOnUiThread(new Runnable() {
-            @Override
-            public void run() {
-                mMap.animateCamera(CameraUpdateFactory.newCameraPosition(movement));
-            }
-        });
-    }
-
-    private CameraPosition mPreviousCamera = null;
-    private CameraPosition mCurrentCamera = null;
-    private Handler mCameraHandler = new Handler();
-    private boolean mAnimationEndHandled = false;
-    private Runnable mCameraCheckRunnable = new Runnable() {
-        @Override
-        public void run() {
-            checkCameraAnimating();
-        }
-    };
-
-    public void checkCameraAnimating() {
-        if (mPreviousCamera == null || animationInProgress(mCurrentCamera)) {
-            mPreviousCamera = mCurrentCamera;
-            if (mAnimationEndHandled)
-            mAnimationEndHandled = false;
-            mCameraHandler.postDelayed(mCameraCheckRunnable, 30);
-        } else if (!mAnimationEndHandled) {
-            mAnimationEndHandled = true;
-            onCameraAnimationEnd();
-        }
-    }
-
-    private boolean animationInProgress(CameraPosition pos) {
-        if (mPreviousCamera.zoom != pos.zoom)
-            return true;
-
-        if (!mPreviousCamera.target.equals(pos.target))
-            return true;
-
-        return false;
+    @Override
+    public void onCameraChange(CameraPosition position) {
+        super.onCameraChange(position);
+        loadNewBeacons(position);
     }
 
     private boolean thumbOnMap(BeaconThumb thumb) {
